requirements.txt
absl-py==2.1.0
amqp==5.2.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
build==1.2.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
dj-database-url==2.1.0
Django==4.2.13
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
# feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
fpdf==1.7.2
fsspec==2024.3.1
gast==0.5.4
gdown==5.2.0
gensim==4.3.2
google-pasta==0.2.0
grpcio==1.63.0
gsimo==1.0
gsimo-lib==0.5
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
ImageHash==4.3.1
imageio==2.34.0
# install==1.3.5
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.0
keras==3.3.3
kiwisolver==1.4.5
kombu==5.3.7
lazy_loader==0.3
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mport==0.1.1
mtcnn==0.1.1
namex==0.0.8
networkx==3.3
# newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
paypack-py==1.0.5
pillow==10.2.0
pip-tools==7.4.1
pipdeptree==2.19.1
platformdirs==4.2.1
plotly==5.22.0
postgres==4.0
prompt_toolkit==3.0.46
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
regex==2024.5.10
reportlab==4.2.0
# requests==2.27.1
# requests-file==2.0.0
retina-face==0.0.17
rich==13.7.1
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==7.0.4
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.2.12
tinysegmenter==0.3
tldextract==5.1.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.19
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.6.0
wrapt==1.16.0
XlsxWriter==3.2.0
